move transformedtargetclassifier up
